404: fix build error - extract canvas into client component
